requirements.txt
aiogram==2.22.1
python-decouple==3.6

